Implemented CharacterParser::evaluatePath
It was a bit of a slog, and I added a helper function for the string
split, because I couldn't get my brain around views::split even though I
think that should be ok.

The tests are passing, apart from the  "Find nemesis with family name"
test. I can't tell if the test is in error, or I've missed something.